+++ tensorflow_advanced_segmentation_models/models/_custom_layers_and_blocks.py
@@ -725,7 +725,7 @@ class GlobalPooling(tf.keras.layers.Layer):
 class MixtureOfSoftMaxACF(tf.keras.layers.Layer):
     def __init__(self, d_k, att_dropout=0.1):
         super(MixtureOfSoftMaxACF, self).__init__()
-        self.temperature = tf.math.pow(tf.cast(d_k, tf.float32), 0.5)
+        self.temperature = tf.math.pow(tf.cast(d_k, self.dtype), 0.5)
         self.att_dropout = att_dropout
         
         self.dropout = tf.keras.layers.Dropout(att_dropout)
